+++ app/src/main/java/com/idroi/marketsense/request/SingleNewsRequest.java
@@ -93,8 +93,8 @@ public class SingleNewsRequest extends Request<CommentAndVote> {
         if(jsonResponse.optBoolean(PARAM_STATUS) && jsonResponse.optJSONObject(PARAM_RESULT) != null) {
             commentAndVote.setRaiseNumber(jsonResponse.optJSONObject(PARAM_RESULT).optInt(PARAM_RAISE_NUMBER));
             commentAndVote.setFallNumber(jsonResponse.optJSONObject(PARAM_RESULT).optInt(PARAM_FALL_NUMBER));
-            commentAndVote.setPrediction(jsonResponse.optJSONObject(PARAM_RESULT).optInt(PARAM_PRED));
-            commentAndVote.setVoting(jsonResponse.optJSONObject(PARAM_RESULT).optInt(PARAM_VOTING));
+            commentAndVote.setPrediction(jsonResponse.optJSONObject(PARAM_RESULT).optDouble(PARAM_PRED));
+            commentAndVote.setVoting(jsonResponse.optJSONObject(PARAM_RESULT).optDouble(PARAM_VOTING));
             if(jsonResponse.optJSONObject(PARAM_RESULT).optJSONArray(PARAM_COMMENTS) != null) {
                 return jsonResponse.optJSONObject(PARAM_RESULT).optJSONArray(PARAM_COMMENTS);
             } else {
